Move CDF printing to StatsAnalysis.py
Outputs a PDF for each simulation containing all of the client's
CDF data in a single plot. That is, sweeping over 4 values of a
parameter will result in 4 plots.
