+++ StatsParser.py
@@ -5,72 +5,6 @@ This class parses the output files of the simulation
 import sys, os, re
 import numpy as np
 import csv
-import matplotlib
-matplotlib.use('Agg')
-import matplotlib.pyplot as plt
-from matplotlib.backends.backend_pdf import PdfPages
-from collections import deque
-from numpy import ma
-from matplotlib import scale as mscale
-from matplotlib import transforms as mtransforms
-from matplotlib.ticker import FixedFormatter, FixedLocator
-import datetime
-
-# Class used to get a nice scale zoomed in on 1 for the y-axis. Code from
-# http://stackoverflow.com/questions/31147893/logarithmic-plot-of-a-cumulative-distribution-function-in-matplotlib
-class CloseToOne(mscale.ScaleBase):
-    name = 'close_to_one'
-
-    def __init__(self, axis, **kwargs):
-        mscale.ScaleBase.__init__(self)
-        self.nines = kwargs.get('nines', 5)
-
-    def get_transform(self):
-        return self.Transform(self.nines)
-
-    def set_default_locators_and_formatters(self, axis):
-        axis.set_major_locator(FixedLocator(
-                np.array([1-10**(-k) for k in range(1+self.nines)])))
-        axis.set_major_formatter(FixedFormatter(
-                [str(1-10**(-k)) for k in range(1+self.nines)]))
-
-
-    def limit_range_for_scale(self, vmin, vmax, minpos):
-        return vmin, min(1 - 10**(-self.nines), vmax)
-
-    class Transform(mtransforms.Transform):
-        input_dims = 1
-        output_dims = 1
-        is_separable = True
-
-        def __init__(self, nines):
-            mtransforms.Transform.__init__(self)
-            self.nines = nines
-
-        def transform_non_affine(self, a):
-            masked = ma.masked_where(a > 1-10**(-1-self.nines), a)
-            if masked.mask.any():
-                return -ma.log10(1-a)
-            else:
-                return -np.log10(1-a)
-
-        def inverted(self):
-            return CloseToOne.InvertedTransform(self.nines)
-
-    class InvertedTransform(mtransforms.Transform):
-        input_dims = 1
-        output_dims = 1
-        is_separable = True
-
-        def __init__(self, nines):
-            mtransforms.Transform.__init__(self)
-            self.nines = nines
-
-        def transform_non_affine(self, a):
-            return 1. - 10**(-a)
-
-        def inverted(self):
-            return CloseToOne.Transform(self.nines)
 
 class StatsParser:
     """
@@ -82,6 +16,7 @@ class StatsParser:
         self.logFiles = []
         self.clientStats = {}
         self.serverStats = {}
+        self.clientCDFData = {}
 
         self.CLIENT = 0
         self.SERVER = 1
@@ -133,8 +68,6 @@ Num Low Priority REQUEST msgs = (?P<numLPReqMsgs>[\d\.]*)
         self.parse_results()
 
     def parse_results(self):
-        mscale.register_scale(CloseToOne)
-
         for subdir, dirs, files in os.walk(self.rootdir):
             for filename in files:
                 fname = os.path.join(subdir, filename)
@@ -145,43 +78,27 @@ Num Low Priority REQUEST msgs = (?P<numLPReqMsgs>[\d\.]*)
                 elif (re.search("Server", filename)):
                     self.parse_stats_log(fname, self.SERVER)
 
-        # Materialize single plot for each similation
-        print "[info] Saving Client CDF."
-        plt.xlabel("Latency (milliseconds)")
-        plt.xscale('log')
-        plt.yscale('close_to_one', nines=4)
-        plt.ylabel("Cumulative Probability")
-        # Decided to not try and print a legend
-        # lgd = plt.legend(bbox_to_anchor=(1.05, 1), loc=2, borderaxespad=0.)
-        plt.title("CDF")
-        plt.grid(True)
-        date = str(datetime.datetime.now())
-        date = '_' + date[:date.find('.')].replace(' ', '_')
-        # TODO: Probably want a PDF output with the same format as the others
-        # pp = PdfPages("./plots/ClientCDF_" + str(self.callCount) + ".pdf")
-        filename = "./plots/ClientCDF_" + date + ".png"
-        plt.savefig(filename) #, bbox_extra_artists=(lgd,), bbox_inches='tight')
-
-        plt.clf() # clear figure for next iteration
-
     def parse_histogram_log(self, filename):
         # To make indexing more readable
         VALUE = 0
         PERCENTILE = 1
         TOTALCOUNT = 2
-        percentiles = deque()
-        latencies = deque()
+        percentiles = []
+        latencies = []
         with open(filename, 'rb') as csvfile:
             reader = csv.reader(csvfile, delimiter=',')
             for row in reader:
+                # Skip the first row
                 if row[VALUE] == "Value":
                     continue
 
-                latencies.append(float(row[VALUE]) / 1000000) #convert to milliseconds
+                latencies.append(float(row[VALUE]))
                 percentiles.append(float(row[PERCENTILE]))
 
-        clientNum = re.search(r'\d+', filename).group(0);
-        plt.plot(latencies, percentiles, label="Client " + str(clientNum))
+        clientNum = re.search(r'\d+', filename).group(0)
+        # Save all of this simulations client's stats in a dictionary with
+        # the client number as the key and the value being a tuple of lists
+        self.clientCDFData[clientNum] = (latencies, percentiles)
 
     def parse_stats_log(self, filename, logType):
         if logType == self.CLIENT:
